move pod2html js to separate file
