+++ root/static/js/cpan.js
@@ -297,117 +297,6 @@ $(document).ready(function() {
             sel.selectedIndex = 0;
         });
     }
-
-    var pod2html_form = $('#metacpan-pod-renderer-form');
-    var pod2html_text = $('[name="pod"]', pod2html_form);
-    var pod2html_update = function(pod) {
-        if (!pod) {
-            pod = pod2html_text.get(0).value;
-        }
-        var submit = pod2html_form.find('input[type="submit"]');
-        submit.attr("disabled", "disabled");
-        var rendered = $('#metacpan-pod-renderer-output');
-        var loading = $('#metacpan-pod-renderer-loading');
-        var error = $('#metacpan-pod-renderer-error');
-        rendered.hide();
-        rendered.html('');
-        loading.show();
-        error.hide();
-        document.title = "Pod Renderer - metacpan.org";
-        $.ajax({
-            url: '/pod2html',
-            method: 'POST',
-            data: {
-                pod: pod
-            },
-            headers: {
-                Accept: "application/json"
-            },
-            success: function(data, stat, req) {
-                var title = data.pod_title;
-                var abstract = data.pod_abstract;
-                document.title = "Pod Renderer - " + title + " - metacpan.org";
-                rendered.html(
-                    '<nav class="toc"><div class="toc-header"><strong>Contents</strong></div>' +
-                    data.pod_index +
-                    '</nav>' +
-                    data.pod_html
-                );
-                var toc = $("nav", rendered);
-                if (toc.length) {
-                    formatTOC(toc[0]);
-                }
-                createAnchors(rendered);
-                loading.hide();
-                error.hide();
-                rendered.show();
-                submit.removeAttr("disabled");
-            },
-            error: function(data, stat) {
-                rendered.hide();
-                loading.hide();
-                error.html('Error rendering POD' +
-                    (data && data.length ? ' - ' + data : ''));
-                error.show();
-                submit.removeAttr("disabled");
-            }
-        });
-    };
-    if (window.FileReader) {
-        $('input[type="file"]', pod2html_form).on('change', function(e) {
-            var files = this.files;
-            for (var i = 0; i < files.length; i++) {
-                var file = files[i];
-                var reader = new FileReader();
-                reader.onload = function(e) {
-                    pod2html_text.get(0).value = e.target.result;
-                    pod2html_update(e.target.result);
-                };
-                reader.readAsText(file);
-            }
-            this.value = null;
-        });
-    }
-    pod2html_form.on('submit', function(e) {
-        e.preventDefault();
-        e.stopPropagation();
-        pod2html_update();
-    });
-
-    var renderer = $(".metacpan-pod-renderer")
-
-    var dragTimer;
-    renderer.on("dragover", function(event) {
-        event.preventDefault();
-        if (dragTimer) {
-            window.clearTimeout(dragTimer);
-        }
-        dragTimer = window.setTimeout(function() {
-            renderer.removeClass("dragging");
-            window.clearTimeout(dragTimer);
-            dragTimer = null;
-        }, 500);
-    });
-
-    $(document).on("dragenter", function(event) {
-        renderer.addClass("dragging");
-    });
-
-    renderer.on("drop", function(event) {
-        event.preventDefault();
-        event.stopPropagation();
-        renderer.removeClass("dragging");
-        if (dragTimer) {
-            window.clearTimeout(dragTimer);
-            dragTimer = null;
-        }
-        var reader = new FileReader();
-        reader.onload = function(e) {
-            pod2html_text.get(0).value = e.target.result;
-            pod2html_update(e.target.result);
-        };
-        reader.readAsText(event.originalEvent.dataTransfer.files[0]);
-    });
 });
 
 function set_page_size(selector, storage_name) {
